oskar_web/requirements.txt
flask==1.0.2
gunicorn==19.9.0
github3.py==1.2.0
